requirements.txt
python-dotenv==0.15.0
terminaltables==3.1.0
numpy==1.20.1
requests==2.25.0